fix: ios runs now

+++ client-data/src/iosMain/kotlin/com/group8/comp2300/util/Sha256.ios.kt
@@ -11,5 +11,5 @@ actual fun sha256(input: String): String {
     val inputBytes = input.toByteArray()
     val digest = UByteArray(32)
     CC_SHA256(inputBytes.refTo(0), inputBytes.size.toUInt(), digest.refTo(0))
-    return digest.joinToString("") { "%02x".format(it) }
+    return digest.joinToString("") { "%02x".format(it.toInt()) }
 }
